refactor SummaryKrsSection and TahunKhs service to utilize centralized fetchData function; improve type definitions and clean up code
