+++ src/types/TahunKhs.d.ts
@@ -1,11 +1,11 @@
 export type TahunKHS = {
-    tahunid: string;
+  tahunid: string;
 };
 
 export type TahunKHSResponse = {
-    status: string;
-    data: TahunKHS[]; // Data dalam bentuk array dalam array
-    code: number;
+  status: string;
+  data: TahunKHS[]; // Data dalam bentuk array dalam array
+  code: number;
 };
 
 type StatusKrs = {
@@ -19,16 +19,18 @@ type StatusKrsResponse = {
   code: number;
 };
 interface JadwalKuliahResponse {
-    status: string;
-    data: {
-      jadwal_id: number;
-      mk_nama: string;
-      hari_id: number;
-      nama_kelas: string;
-      jam_mulai: string;
-      jam_selesai: string;
-      ruang_id: string;
-      dsn: string;
-    }[];
-    code: number;
-  }
+  status: string;
+  data: JadwalKuliah[][];
+  code: number;
+}
+
+interface JadwalKuliah {
+  jadwal_id: number;
+  mk_nama: string;
+  hari_id: number;
+  nama_kelas: string;
+  jam_mulai: string;
+  jam_selesai: string;
+  ruang_id: string;
+  dsn: string;
+}
